Rename xarray.py -> xarray_support.py.
Otherwise, can't import xarray global package inside our module.
