+++ siphon/cdmr/tests/test_xarray.py
@@ -2,7 +2,7 @@
 # Distributed under the terms of the MIT License.
 # SPDX-License-Identifier: MIT
 
-from siphon.cdmr.xarray import CDMRemoteStore
+from siphon.cdmr.xarray_support import CDMRemoteStore
 from siphon.testing import get_recorder
 from numpy.testing import assert_almost_equal
 from xarray import open_dataset
